[Refactor] Главная страница вынесена в /pages
