+++ src/components/burger-ingredients/burger-ingredients.module.css
@@ -43,15 +43,6 @@
   height: 100%;
 }
 
-.loading {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  flex-flow: column nowrap;
-  justify-content: center;
-  align-items: center;
-}
-
 .drop {
   position: relative;
   display: flex;
