+++ src/pages/index.js
@@ -1,0 +1,3 @@
+import MainPage from './main-page/main-page';
+
+export { MainPage };
