+++ src/pages/main-page/main-page.module.css
@@ -1,0 +1,16 @@
+.main {
+  display: grid;
+  grid-template-columns: repeat(2, 600px);
+  gap: 0 40px;
+  overflow: hidden;
+  height: 100vh;
+}
+
+.loading {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-flow: column nowrap;
+  justify-content: center;
+  align-items: center;
+}
